Add ogg2gme.py wrapper script around tttool.
Outline:
ogg2gme.py is a Python script which allows users which are
unfamiliar with tttool, oids and the *.yaml files to produce
*.gme files and *.png files containing OID labels. All one
needs to do is:
- record *.ogg files into a specifically named directory
- double click a *.bat file on Windows (or invoke ogg2gme.py on Linux)
- copy the generated *.gme file onto the stick
- print the OID labels (*.png file)

The script produces a simple 1:1 mapping from *.ogg files to OIDs.
This is sufficient to add Tip-Toi support to non-Tip-Toi books and
to produce speaking labels for anything.

The generated *.yaml file can be manually modified any time (for
example using more complicated mechanisms like registers). The
modifications are preserved across ogg2gme.py invocations.

See ogg2gme/README.txt for details (German).